+++ mu4/cloud/CMakeLists.txt
@@ -26,12 +26,13 @@ set(MODULE_QML_IMPORT ${CMAKE_CURRENT_LIST_DIR}/qml)
 set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/cloudmodule.cpp
     ${CMAKE_CURRENT_LIST_DIR}/cloudmodule.h
-    ${CMAKE_CURRENT_LIST_DIR}/controllers/iaccountcontroller.h
-    ${CMAKE_CURRENT_LIST_DIR}/controllers/accountcontroller.h
-    ${CMAKE_CURRENT_LIST_DIR}/controllers/accountcontroller.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/iaccountcontroller.h
+    ${CMAKE_CURRENT_LIST_DIR}/imp3exporter.h
+    ${CMAKE_CURRENT_LIST_DIR}/cloudtypes.h
     ${CMAKE_CURRENT_LIST_DIR}/view/accountmodel.h
     ${CMAKE_CURRENT_LIST_DIR}/view/accountmodel.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/internal/imp3exporter.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/accountcontroller.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/accountcontroller.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/cloudmanager.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/cloudmanager.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/cloudmanager_p.h
